Use `memmap2` for accessing file contents
This is done as a replacement for the code that
loaded the entire contents of the log files to memory.
This way, we don't have to handle this logic ourselves +
we get to be more memory efficient, because the kernel will use
physical pages only when necessary.

Note that we do need to use `unsafe` here, and that changes to the
`mmap`ed files are UB.
This has to be solved, with an OS specific solution later, for now,
rnav just doesn't support file updates.